Added Protected Routes based on the User and Admin in the Projects Route.

+++ backend/src/project/project.controller.ts
@@ -1,7 +1,10 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Patch, Post, Put, Request, UnauthorizedException, UseGuards } from '@nestjs/common';
 import { ProjectService } from './project.service';
 import { Project } from "./project.schema";
 import { CreateProjectDto } from './create-project.dto';
+import { Roles } from 'src/auth/user/roles.decorator';
+import { UserRole } from 'src/models/users.models';
+import { AuthGuard } from 'src/auth/auth.guard';
 
 @Controller('project')
 export class ProjectController {
@@ -9,26 +12,67 @@ export class ProjectController {
 
 
     @Get()
-    async getProjects(): Promise<Project[]> {
-        return this.projectService.findAll();
+    @UseGuards(AuthGuard)
+    async getProjects(@Request() req): Promise<Project[] | Project> {
+        const user = req.user;
+
+        if(user.role === UserRole.ADMIN) {
+            // ADMIN: Show all the projects
+            return this.projectService.findAll();
+        }else if (user.role === UserRole.USER) {
+            // Regular User: Show only thier Project.
+            return this.projectService.findParticularProject(user.userId);
+        }else {
+            throw new UnauthorizedException("Not Authorized to view projects");
+        }
+
     }
+    
     @Get("/:id")
     async getParticularProject(@Param('id') id: string) : Promise<Project> {
         return this.projectService.findParticularProject(id);
     }
 
     @Post() 
-    async createProject(@Body() createProjectDto: CreateProjectDto): Promise<Project> {
-        return this.projectService.create(createProjectDto);
+    @UseGuards(AuthGuard)
+    async createProject(@Body() createProjectDto: CreateProjectDto, @Request() req): Promise<Project> {
+        const userId = req.user.userId;
+
+        return this.projectService.create(createProjectDto, userId);
     }
+
     @Delete("/:id")
-    async deleteProject(@Param('id') id: string): Promise<Project> {
-        return this.projectService.deleteProject(id);
+    @UseGuards(AuthGuard)
+    async deleteProject(@Param('id') id: string, @Request() req): Promise<Project> {
+        const user = req.user;
+
+        if (user.role === UserRole.ADMIN) {
+            return this.projectService.deleteProject(id);
+        } else {
+            const project = await this.projectService.findParticularProject(id);
+            if (project.ownerId.toString() === user.userId) {
+                return this.projectService.deleteProject(id);
+            } else {
+                throw new UnauthorizedException("Not Authorized to delete this project");
+            }
+        }
     }
-    // @Patch()
+
     @Put("/:id")
-    async updateProject(@Param('id') id: string,  @Body() updateData: Partial<Project>): Promise<Project> {
-        return this.projectService.updateProject(id, updateData);
+    @UseGuards(AuthGuard)
+    async updateProject(@Param('id') id: string, @Body() updateData: Partial<Project>, @Request() req): Promise<Project> {
+        const user = req.user;
+
+        if (user.role === UserRole.ADMIN) {
+            return this.projectService.updateProject(id, updateData);
+        } else {
+            const project = await this.projectService.findParticularProject(id);
+            if (project.ownerId.toString() === user.userId) {
+                return this.projectService.updateProject(id, updateData);
+            } else {
+                throw new UnauthorizedException("Not Authorized to update this project");
+            }
+        }
     }
 
 }
